Refactor name styling
