+++ src/pages/IndicatorsManagement.jsx
