+++ src/pages/NewIndicator.jsx
